requirements.txt
asgiref==3.8.1
Django==5.0.4
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
django-solo==2.2.0
django-admin-sortable2==2.1.10
django-tinymce==4.0.0
isbnlib==3.10.14
django-object-actions==4.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
django_fieldsets_with_inlines @ git+https://github.com/joaoseckler/django-fieldsets-with-inlines/@bbb927175947bbef4b762d20289065b922c9c56e
lxml==5.2.1
soupsieve==2.5
Unidecode==1.3.8
pillow==10.3.0
django-colorfield==0.11.0
defusedxml==0.7.1
fonttools==4.51.0
fpdf2==2.7.8
python-barcode==0.15.1
djangorestframework==3.15.1
djangorestframework-csv==3.0.2
